requirements.txt
# Fork of gym with more flexible utility functions.
gym @ git+https://www.github.com/openai/gym@8819d561132082f6130d4a2388c68a963f41ec4f#egg=gym
# nngeometry module used in the EWC method
nngeometry @ git+https://github.com/oleksost/nngeometry.git#egg=nngeometry
# Temporary fix for issue#128
pyyaml!=5.4.*,>=5.1
simple_parsing>=0.0.16.post2
# matplotlib==3.2.2
matplotlib
# NOTE: @lebrice: PyTorch suddenly got really picky about type annotations in 1.9.0 for
# some reason, and they really don't do a great job at evaluating them, so removing it
# for now.
torch==1.8.1
torchvision==0.9.1
scikit-learn
tqdm
continuum==1.0.19
# Only required for the current demo:
wandb
plotly
pandas
# Only for python < 3.8
singledispatchmethod;python_version<'3.8'
# NOTE: PyTorch-Lightning version 1.4.0 is "working" but raises lots of warnings.
pytorch-lightning==1.5.9
lightning-bolts==0.5.0
# Requirements for running tests:
pytest-timeout
pytest-xdist
pytest-xvfb # Prevents the gym popups from displaying during tests.
# Required for the RL methods
# TODO: Move this to an 'extras' in setup.py?
stable-baselines3>=0.11.1
pyvirtualdisplay
# Required for the synbols dataset to work. 
h5py
